requirements.txt
Django==2.1
robotframework==3.0.4
djangorestframework-jwt==1.11.0
djangorestframework == 3.8.2
Markdown==2.6.11
django-filter==2.0.0
mysqlclient==1.3.12
requests==2.19.1
redis==2.10.6